Revert HeaderMod and move to index.ts
